--- lib/partidas/timer.dart
@@ -1,56 +1,0 @@
-/*
-
-import 'package:flutter/material.dart';
-import 'package:simple_timer/simple_timer.dart';
-
-const COUNT_DOWN_SEGS = 30;
-
-Color purpleColor = Color(0xff6A1B9A);
-
-class TimerTemplate extends StatefulWidget {
-  final width;
-  final height;
-  final turnOk;
-
-  TimerTemplate({this.width, this.height, this.turnOk});
-
-  @override
-  _TimerTemplate createState() => _TimerTemplate(
-      width: this.width, height: this.height, turnOk: this.turnOk);
-}
-
-class _TimerTemplate extends State<TimerTemplate>
-     {
-  final width;
-  final height;
-  final turnOk;
-
-
-
-  _TimerTemplate({this.width, this.height, this.turnOk});
-
-
-
-
-
-  @override
-  Widget build(BuildContext context) {
-    if (this.turnOk == true) {
-      return
-    }
-    return Container(
-      height: this.height,
-      width: this.width,
-      color: Color(0xffC5C5C5),
-      alignment: Alignment.center,
-      child: Text(
-        "0",
-        style: TextStyle(
-            color: Colors.white,
-            fontWeight: FontWeight.bold,
-            fontSize: this.width * 0.2),
-      ),
-    );
-  }
-}
-*/
